Saving completion time in hours

+++ final/app/src/main/java/com/example/makeitso/model/Task.kt
@@ -29,6 +29,6 @@ data class Task(
   val url: String = "",
   val flag: Boolean = false,
   val completed: Boolean = false,
-  val completionTime: Int? = null,
+  val completionTime: Long? = null,
   val userId: String = ""
 )
